streamlit_app/requirements.txt
streamlit>=1.30
kafka-python>=2.0.2
pandas>=2.2
python-dateutil>=2.8
